llm: add an ollama image puller
now we can pull ollama image from the internet and run it like this:

pull and run the image:
  $ ./wk run hf-mirror.com/bartowski/Llama-3.2-1B-Instruct-GGUF:Q8_0

pull the image only:
  $ ./wk pull hf-mirror.com/bartowski/Llama-3.2-1B-Instruct-GGUF (default tag is Q8_0)

+++ llm/parser.go
@@ -158,7 +158,7 @@ func GGUFParser(filename string) (*GGUFFile, error) {
 		return nil, fmt.Errorf("failed to read header: %w", err)
 	}
 
-	pbar.Max = float64(gguf.Header.KVCount + gguf.Header.TensorCount*2 + 3)
+	pbar.Max = int64(gguf.Header.KVCount + gguf.Header.TensorCount*2 + 3)
 
 	if string(gguf.Header.Magic[:]) != "GGUF" {
 		return nil, fmt.Errorf("invalid magic number: expected 'GGUF', got '%s'", string(gguf.Header.Magic[:]))
